server/requirements.txt
flask
flask-cors
python-dotenv
SQLAlchemy 
psycopg2-binary 
Flask-Bcrypt